git-svn-id: https://svn.apache.org/repos/asf/sis/branches/JDK6@1697310 13f79535-47bb-0310-9956-ffa450edef68

+++ core/sis-utility/src/main/java/org/apache/sis/measure/NumberRange.java
@@ -208,7 +208,7 @@ public class NumberRange<E extends Number & Comparable<? super E>> extends Range
         if (Float.isNaN(value)) {
             throw new IllegalArgumentException(Errors.format(Errors.Keys.NotANumber_1, name));
         }
-        return (value != infinity) ? Float.valueOf(value) : null;
+        return (value != infinity) ? value : null;
     }
 
     /**
